apps/users/requirements.txt
fastapi[standard]
python-jose[cryptography]
passlib[bcrypt]
tortoise-orm[asyncmy]
aerich[toml]
aiofiles